Refactor files for component like structure
