+++ app/src/js/modules/utilities.js
@@ -1,0 +1,49 @@
+export function addGlobalEventListener(type, selector, callback, parent = document) {
+  parent.addEventListener(type, e => {
+    if (e.target.matches(selector)) {
+      callback(e);
+    }
+  });
+}
+
+export function getPageRange(currentPage, totalPages, visibleRange) {
+  const range = [];
+  const start = Math.max(2, currentPage - visibleRange);
+  const end = Math.min(totalPages - 1, currentPage + visibleRange);
+
+  range.push(1);
+
+  if (start > 2) range.push('...');
+
+  for (let i = start; i <= end; i++) {
+    range.push(i);
+  }
+
+  if (end < totalPages - 1) range.push('...');
+
+  if (totalPages > 1) range.push(totalPages); 
+
+  return range;
+}
+
+export function updateHeadingVisibility() {
+  const projectSections = document.querySelectorAll(".project_section");
+  projectSections.forEach(section => {
+    let hasVisibleCard = false;
+    // Select the report heading within the current section
+    let reportHeading = section.querySelector('.report_heading');
+    
+    // Select all project cards within the current section
+    let projectCards = section.querySelectorAll('.project-card');
+    
+    // Flag visible card if any card is not hidden
+    projectCards.forEach((card) => {
+      if (!card.hidden) {
+        hasVisibleCard = true;
+      }
+    })
+    
+  // Hide heading if all cards under section are hidden
+  reportHeading.hidden = !hasVisibleCard;
+  });
+}
