+++ package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "@types/dompurify": "^3.0.5"
+  }
+}
